requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
psycopg==3.2.2
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
